fix(terminal): match agent keyboard press feedback
Share the gray press highlight, scale, release timing, and light tap haptics across Agent and Terminal keys while retaining modifier selection states.

refs #302
